hwc2: handle create core interface error condition
In some scenarios when call to "drmOpen("msm_drm", nullptr)"
is failed due to any reason then the coreintf_ is not created
and it will remain NULL.

The error returned from coreinterface class is not handled
in HWC Session class. As a result when coreintf_ is dereferenced
to make any calls then the composer process crashes with a NULL
pointer exception.

Change is to handle the error condition in HWC Session and return
the appropriate error back to HWCSession::Init().

Change-Id: If9eabb115042a904f226d09c06e35ee438493201
Signed-off-by: Rahul Sharma <rahsha@codeaurora.org>