gralloc: Consider custom metadata region size when freeing buffer
Buffer release logic currently does not consider the size of the dynamically
allocated custom content metadata region. This results in memory leaks when
the region is present.

Update buffer unmap logic to retrieve the size of the custom content metadata
region and factor it into the size of the memory region being freed.

Change-Id: I277b8c249ed037fd5ccbc085df9b4d72bf41b563