sdm: skip first commit check for poweroff
When multiple displays are enabled with non-mirror mode
and non primary displays are not rendering any content the
display is not powered off due to the first commit check in the
poweroff function. Due to this the STR use case is failed
as dispay on which no contents are rendered are not powered off.
The issue is resolved by removing the first commit check in the
poweroff function.

Change-Id: I702d4004a48cac03f1054a73a97a5b2b5524bb48
Acked-by: Praveen Kanchikatte Prabhu <c_kanchi@qti.qualcomm.com>
Signed-off-by: Rahul Sharma <quic_rahsha@quicinc.com>
(cherry picked from commit 68a549348db84e26ceb32515fee29bcb0475315e)