composer: set panel brightness through DRM property
For HDR use cases where brightness can change with each frame,
the conventional way of setting brightness through sysfs node
may lead to mismatch in the update of brightness value and the
display of corresponding frame due to delay in commit. Thus,
the out of sync update of brightness can appear as a flicker.

The change is to program the brightness value for the current
frame to the driver as part of same commit cycle.

Change-Id: I3047ea38915210a2ea1c9f58ff81712e2cc81d21
(cherry picked from commit a28f21296f94962714bbc4dad6ca2bee417c7053)