gralloc: Fix size for memcpy of dynamic metadata
The size of std::optional was being used rather than the size of
the underlying vector. This impacts the standard metadata set
path, leading to a segfault if a size 0 vector is passed in,
or only copying the first 32 bytes of the input vector.

Change-Id: I64f20683041cec4c2a0941cb8a58f1b51dd92883
(cherry picked from commit 408da7afe2e51e7320f4f3314d41f00321f3fb65)