sdm : fix the incorrect drm card path for msm_drm
In case of Dual DPU when multiple msm_drm cards are parsed, for DPU1
path_ variable is overwritten since the iteration continous for other
cards.
Fix : Adding local variable for loop iteration and copying the card path
in case it matches the card having "msm_drm"
This fixes hpd issue on DP ports.

Change-Id: Ieef80f3843849240bd5ff985c7921021406b68d2