sdm: hwc2: change hwc_session type to auto
Change the hwc_session type from hard pointer to auto type.
A race condition between hwc and graphics driver can result
in the hwc trying to delete a HWCSession object. The
HWCSession object is converted to a strong pointer without
a proper owner, and trying to delete it may lead to a crash.

Signed-off-by: Yashovardhan Bhatt <quic_ybhatt@quicinc.com>
Change-Id: Ic308847bceb9b32a127a6c78828d236dc4e82411