sdm: updated card opening logic for more than 2 DPUs
DRMMaster::Init was not returning a failure in case of more than 2
DPUs and was always opening card for DPU1 for all the DPU instances
(from 1 to 15). Always returning fd for more than available DPUs is not
correct, updated the logic to return correct fd for only available DPU
instances.

Change-Id: I1eb2960d07d2ad6e89197ec3c0f4e2d36777b5fb
Signed-off-by: Tushar Shivajirao Ghadge <quic_tghadge@quicinc.com>