init: Enable brightness_drm_prop by default in the mainline
This change enables sending of the panel brightness through a DRM
property during the commit.

composer: Detect commit for the setDisplayBrightness command

This change identifies if the current draw cycle is triggering a
commit. This determines if the brightness update is going to be
updated in the sysfs node or the drm property.

Change-Id: Ice70e4a8aea90cab79f9dbc457f01183e759ef0a
CRs-Fixed: 4107440