sdm: Retry HDCP level change with primary display
As part of customer configuration, disp_id was hard coded with primary
display(Built-in display),
which caused issue in default configuration as HDCP level cannot be
applied on built_in display.

Fix: Fix has been made considering both default and customer
configuration.
Attempt HDCP level change on the requested display ID.
If the requested display is not connected and is not the primary, it
retries with the primary display.
If the primary display is also not connected, it returns an error.

Change-Id: I35cecd9e005016607a779a04eb99e7b2719a998b
Signed-off-by: Geetanjali Patil <geetanja@qti.qualcomm.com>