display: config: Cleanup hdmirotationlock property
Remove setting of persist.demo.hdmirotationlock in
display-product.mk as this is causing AVC denials at
bootup time.

This property is not needed in automotive projects.
Also, this was getting set to false, which will be the usual
default value if the property is not set.
So, avoid setting this property.

Change-Id: Iaf7005df7c47eabe3b7fb0750e935b19db7cc255